add model add, update and delete task

+++ cli/cmds/remove_task.py
@@ -9,11 +9,14 @@ class DeleteTask(AbsCommand):
         try:
             task_id = int(task_id)
             task = self.dashboard.get_task(task_id)
-            print('are you sure man?')
         except ValueError as err:
             print(f'Incorrect value {task_id}')
         except IndexError as err:
             print(f'No task with this id, {task_id}')
         else:
-            self.dashboard.delete_task(task_id)
-
+            decision_input = input('are you sure man? Write "yes": ')
+            if decision_input == str('yes'):
+                self.dashboard.delete_task(task_id)
+                print(f'I deleted task number {task_id}')
+            else:
+                print('Sorry, i can not delete it.')
